Use jekyll to generate some static content
This provides the infrastructure for generating static pages using jekyll,
for convenient templating and other features.

This commit defines jekyll's includes and layouts, and generates
the site's css, js, version-controlled images, and the toplevel index.
More content will be converted later.

A README.md has been added describing how jekyll is used for this site.

The only content changes in the toplevel index are:
* The nonfunctioning "ABOUT" link has been removed from the menu bar.
* The link to wiki has been moved from the page footer to the menu bar.
* The "[if IE]" rules have been removed.