Update ClusterLabs logo
Many thanks to designer Kristoffer Grönlund <kgronlund@suse.com>!
This replaces the site icons and the wiki logo.

The wiki logo is not the best adaptation -- it's just the original with a
transparent background -- but this helps get the new logo out while we come up
with a better approach. Maybe we could tweak the logo's color scheme to better
match the wiki banner background, or tweak the background to better match the
new logo, and add "ClusterLabs" text to the banner.